+++ extension/package.json
@@ -21,37 +21,22 @@
       {
         "command": "antfu.file-nesting.manualUpdate",
         "title": "Update config now",
-        "category": "File Nesting"
+        "category": "File Nesting Updater"
       }
     ],
     "configuration": {
       "type": "object",
       "title": "File Nesting Updater",
       "properties": {
-        "fileNestingUpdater": {
-          "type": "object",
-          "properties": {
-            "enabled": {
-              "type": "boolean",
-              "description": "Enable file nesting",
-              "default": true
-            },
-            "maxDepth": {
-              "type": "number",
-              "description": "Maximum nesting depth",
-              "default": 3
-            },
-            "maxFiles": {
-              "type": "number",
-              "description": "Maximum files per folder",
-              "default": 10
-            },
-            "maxSize": {
-              "type": "number",
-              "description": "Maximum size per folder",
-              "default": 10
-            }
-          }
+        "fileNestingUpdater.autoUpdate": {
+          "type": "boolean",
+          "description": "Fetch and update the latest config automatically",
+          "default": true
+        },
+        "fileNestingUpdater.promptOnAutoUpdate": {
+          "type": "boolean",
+          "description": "Should show up the prompt before doing auto update",
+          "default": true
         }
       }
     }
