+++ extension/src/constants.ts
@@ -1,0 +1,2 @@
+export const URL = 'https://cdn.jsdelivr.net/gh/antfu/vscode-file-nesting-config@main/README.md'
+export const MSG_PREFIX = 'File Nesting Updater:'
